% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trace.R
\name{tidy_raw_timedots}
\alias{tidy_raw_timedots}
\title{Correct raw time dots csv file from ImageJ}
\usage{
tidy_raw_timedots(filepath_timedots, col_x = 1, col_y = 2)
}
\arguments{
\item{filepath_timedots}{the path to the csv file containing the raw time dots
data from ImageJ}

\item{col_x}{column number of the x values, default is set to first column.}

\item{col_y}{column number of the y values, default is set to second column.}
}
\value{
One trace data frame after correcting ImageJ's default origin
placement
}
\description{
This function takes the raw time dots csv file from ImageJ and applies a
series of corrections to it such that it can be used in this package. This
involves correcting the default y-axis values, ordering, filtering, and
adding proper names to the data frame.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "WS_25_1981", package = "recoverKBTDR")
filepath_timedots <- paste(filepath, "WS_25_1981_time_dots.csv", sep = "/")
tidy_raw_timedots(filepath_timedots, col_x = 1, col_y = 2)
}

}
